base: Add Utils.foregroundList()
Add foregroundList() in Utils that returns a list of surface items
currently in foreground.
At the moment in the compositor base, ForegroundAppInfoMgr is the only
use case of this method. But it can also be useful for product
compositor if it wants to do something based on what is currently in
foreground.
Architecturally it is intended for this method to get a list of items as
an argument and return a result based on it instead of relying on
"views" which can be used globally. It is because that the method is
implemented in a singleton object Utils which is desirable not to depend
on any other QML object if possible.

Change-Id: I84b9bdefbe9f0d86883328556d6cc0168ccca4bc