Set permission of wayland socket
Set permission ug+rw to the wayland socket created and inherit ownership
from the parent directory.
It had been done in the startup script but this is a better place to do
that since it can be done right after the wayland socket is created.
The parent directory will be created in the start-up script as before
and the ownership will be set too.

Change-Id: Iecca3bdc03cca9249507ce4a1a99eba06440c164